requirements.txt
annotated-types==0.7.0
idna==3.10
oauthlib==3.2.2
openai==1.61.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0

# Core Google API requirements
google-api-python-client>=2.108.0
google-api-core==2.24.1
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0


# Python Telegram Bot 
python-telegram-bot==21.10
